+++ page-texture.php
@@ -1,46 +1,38 @@
-<?php get_header(); ?>
+<?php
+/**
+ * Template Name: Texture Page
+ */
+/*
+ * Generate a unique ID for each form and a string containing an aria-label
+ * if one was passed to get_search_form() in the args array.
+ */
+get_header();
+?>
 <div class="projects">
 	<?php
-		// these are the two posts that belong to this page
-		//13 = <<page_home></page_home></page_home>
+		get_template_part( 'template-parts/loops/top-posts', NULL,  array('page-name' => 'page-texture') );
+		// get template part projects
+		// get all the posts, except the ones with category name "hidden"
 
-		$page_posts = get_posts( array('category_name' => 'page-texture') );
-		$hover_text = get_field('hover_text', $post->ID);
-		foreach($page_posts as $post):
-			$block = parse_blocks( $post->post_content );
-			$hover_text = $block[0]['attrs']['data'];
+			$all_posts = get_posts(array(
+					'post_type' => 'post',
+					'category__not_in' => array(get_cat_ID('hidden')),
+					'posts_per_page'   => -1,
+				)
+		);
+
+		if (count($all_posts) > 1):
+			foreach($all_posts as $post):
 		?>
-
-			<h2 class="project-line top-menu-line">
-				<a href="<?php the_permalink(); ?>"><?php echo $hover_text['project_hover_text']; ?></a>
+			<h2 class="project-line">
+				<a href="<?php the_permalink(); ?>"
+					title="<?php the_title(); ?>"
+					alt="Project of Tabea Nixdorff: <?php the_title(); ?>"
+				>
+				<?php the_title(); ?>
+				<?php parse_ACF_block( 'acf/project' ); ?>
+				</a>
 			</h2>
-		<?php endforeach; ?>
-	<?php
-		// this is filtered by pre-get-posts
-		if (have_posts()): while (have_posts()) : the_post();
-		$posttags = get_the_tags();
-	?>
-    	<h2 class="project-line">
-			<a href="<?php the_permalink(); ?>
-				"title="<?php the_title(); ?>"
-				alt="Project of Tabea Nixdorff: <?php the_title(); ?>"
-				data-tag="<?php if ($posttags):
-					foreach($posttags as $tag): echo $tag->name . ' ';
-					endforeach; endif;?>"
-			>
-			<?php the_title(); ?>
-			<?php parse_ACF_block( 'acf/project' ); ?>
-			</a>
-		</h2>
-  	<?php endwhile;
-	else:
-		echo 'no projects right now'; ?>
-
-	<?php endif; ?>
-	<div class="leftover-projects"></div>
-</div>
-
-
-<?php
+	<?php endforeach; endif;
 	get_footer();
 ?>